Support reverse mounting for IMUs (uint8_t version)
- Added `reversed` property to `board-defaults.json` schema.
- Added `PRIMARY_IMU_MOUNTING_REVERSE` and `SECONDARY_IMU_MOUNTING_REVERSE` macros in `src/defines.h`.
- Updated `scripts/preprocessor.py` to pass the `reversed` flag to `SENSOR_DESC_ENTRY`.
- Updated `src/boards/boards_default.h` to use mounting reverse macros.
- Updated `BNO080Sensor` and `SoftFusionSensor` to handle `extraParam` with `MOUNTING_REVERSE` flag, keeping `extraParam` as `uint8_t`.
- Updated `SensorBuilder` to use `uint8_t` for `extraParam`.

+++ src/sensors/bno080sensor.h
@@ -43,7 +43,7 @@ public:
 		float rotation,
 		SlimeVR::SensorInterface* sensorInterface,
 		PinInterface* intPin,
-		int extraParam
+		uint8_t extraParam
 	)
 		: Sensor(
 			"BNO080Sensor",
@@ -86,7 +86,7 @@ protected:
 		float rotation,
 		SlimeVR::SensorInterface* sensorInterface,
 		PinInterface* intPin,
-		int extraParam
+		uint8_t extraParam
 	)
 		: Sensor(sensorName, imuId, id, registerInterface, rotation, sensorInterface)
 		, m_IntPin(intPin) {
